+++ src/main.rs
@@ -7,15 +7,13 @@ mod m3u_filter_error;
 mod config_reader;
 mod model;
 mod filter;
-mod m3u_parser;
-mod playlist_processor;
 mod repository;
 mod download;
 mod utils;
 mod messaging;
-mod xtream_parser;
 mod test;
 mod api;
+mod processing;
 
 use env_logger::{Builder};
 use log::{debug, error, info, LevelFilter};
@@ -25,6 +23,7 @@ use crate::config_reader::{read_api_proxy_config, read_config, read_mappings};
 use crate::messaging::send_message;
 use crate::model::config::{Config, ProcessTargets, validate_targets};
 use crate::m3u_filter_error::{M3uFilterErrorKind};
+use crate::processing::playlist_processor;
 
 #[derive(Parser)]
 #[command(name = "m3u-filter")]