+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added Xtream Api Endpoints.
 * Added multiple input support
 * Added Telegram message support
+* Added Target watch for groups
 
 # v1.0.1(2023-09-07)
 * Refactored sorting. Sorting channels inside group now possible